Move TimeseriaAnomalyDetector in base.py
