+++ ats/anomaly_detectors/base.py
@@ -2,11 +2,12 @@
 """Anomaly detectors"""
 
 import pandas as pd
+from ats.utils import convert_timeseries_df_to_timeseries, convert_timeseries_to_timeseries_df
+from timeseria.datastructures import TimeSeries
 
 # Setup logging
 import logging
 logger = logging.getLogger(__name__)
-
 
 class AnomalyDetector():
 
@@ -34,3 +35,45 @@ class AnomalyDetector():
         """
         raise NotImplementedError()
 
+class TimeseriaAnomalyDetector(AnomalyDetector):
+    """
+    Base class for anomaly detectors wrapped from the timeseria library.
+    """  
+    
+    model_class = None
+
+    def __init__(self, *args, **kwargs):
+        if self.model_class is None:
+            raise NotImplementedError('Subclasses must define a timeseria model')
+        self.model = self.model_class(*args, **kwargs)
+
+    def fit(self, data, *args, **kwargs):
+        """
+        Fit the timeseria anomaly detector model.
+        """      
+        if not isinstance(data,pd.DataFrame):
+            raise NotImplementedError('Not yet implemented for non DataFrame inputs')
+        timeseries_df = data
+
+        # Using timeseria to fit the model
+        timeseries = convert_timeseries_df_to_timeseries(timeseries_df)
+        model = self.model
+        model.fit(timeseries, *args, **kwargs)
+
+    def apply(self, data, *args, **kwargs):
+        """
+        Apply the timeseria anomaly detector model.
+        """     
+        if not isinstance(data,pd.DataFrame):
+            raise NotImplementedError('Not yet implemented for non DataFrame inputs')
+        timeseries_df = data
+
+        # Using timeseria to fit and apply the model
+        timeseries = convert_timeseries_df_to_timeseries(timeseries_df)
+        timeseries = self.model.apply(timeseries, *args, **kwargs)
+
+        # Convert back to DataFrame
+        timeseries_df = convert_timeseries_to_timeseries_df(timeseries)
+        timeseries_df['anomaly'] = (timeseries_df['anomaly'].astype(float) != 0).astype(int)
+
+        return timeseries_df
